+++ lib/sfio/sfvscanf.c
@@ -21,7 +21,7 @@
 **	Written by Kiem-Phong Vo.
 */
 
-#define MAXWIDTH	(int)(((uint)~0)>>1)	/* max amount to scan   */
+#define MAXWIDTH	INT_MAX	// max amount to scan
 
 /**
  * @param form format string
@@ -575,7 +575,7 @@ int sfvscanf(FILE *f, const char *form, va_list args)
 		    if (fmt == 'd' || fmt == 'i')
 			*((int *) value) = (int) argv.ll;
 		    else
-			*((uint *) value) = (uint) argv.lu;
+			*((unsigned*)value) = (unsigned)argv.lu;
 		}
 	    }
 	} else if (fmt == 's' || fmt == 'c' || fmt == '[') {